Réorganise le code d’import des sources de données
